refactor: wallet apis
