--- packages/maskbook/src/extension/options-page/DashboardComponents/TransactionList/types.ts
@@ -1,1 +1,0 @@
-export type { Transaction } from '../../../../plugins/Wallet/apis'
