+++ packages/maskbook/src/plugins/Wallet/services/index.ts
@@ -1,3 +1,4 @@
 export * from './wallet'
 export * from './token'
-export * from '../apis'
+export * from './assets'
+export * from './transactions'
